修正插件系统加载部分
