+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/defaultplugin/zaicomic/json/SearchItem.kt
@@ -1,0 +1,8 @@
+package indi.dmzz_yyhyy.lightnovelreader.defaultplugin.zaicomic.json
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchItem(
+    @SerializedName("id")
+    val id: Int
+)
