+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/defaultplugin/zaicomic/json/ZaiComicData.kt
@@ -1,0 +1,22 @@
+package indi.dmzz_yyhyy.lightnovelreader.defaultplugin.zaicomic.json
+
+import com.google.gson.annotations.SerializedName
+
+data class ZaiComicData<T>(
+    @SerializedName("errno")
+    val errno: Int,
+    @SerializedName("errmsg")
+    val errmsg: String,
+    @SerializedName("data")
+    val data: T
+)
+
+data class DataContent<T>(
+    @SerializedName("data")
+    val data: T
+)
+
+data class ListDataContent<T>(
+    @SerializedName("list")
+    val list: List<T>?
+)
