--- plugin/build.gradle.kts
@@ -1,3 +1,0 @@
-plugins {
-    kotlin("jvm")
-}
