--- plugin/defaultdatasource/src/main/kotlin/io/nightfish/defaultdatasource/zaicomic/json/RecommendData.kt
@@ -1,21 +1,0 @@
-package io.nightfish.defaultdatasource.zaicomic.json
-
-import com.google.gson.annotations.SerializedName
-
-data class RecommendData (
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("data")
-    val data: List<RecommendItem>
-)
-
-data class RecommendItem (
-    @SerializedName("obj_id")
-    val id: Int,
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("cover")
-    val cover: String,
-    @SerializedName("type")
-    val type: Int
-)
