--- plugin/defaultdatasource/src/main/kotlin/io/nightfish/defaultdatasource/zaicomic/json/SearchItem.kt
@@ -1,8 +1,0 @@
-package io.nightfish.defaultdatasource.zaicomic.json
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchItem(
-    @SerializedName("id")
-    val id: Int
-)
