--- plugin/defaultdatasource/src/main/kotlin/io/nightfish/defaultdatasource/zaicomic/json/TagTypeItem.kt
@@ -1,10 +1,0 @@
-package io.nightfish.defaultdatasource.zaicomic.json
-
-import com.google.gson.annotations.SerializedName
-
-data class TagTypeItem(
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("tagId")
-    val tagId: Int
-)
