--- plugin/defaultdatasource/src/main/kotlin/io/nightfish/defaultdatasource/zaicomic/json/UpdatePageItem.kt
@@ -1,12 +1,0 @@
-package io.nightfish.defaultdatasource.zaicomic.json
-
-import com.google.gson.annotations.SerializedName
-
-data class UpdatePageItem(
-    @SerializedName("comic_id")
-    val id: Int,
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("cover")
-    val cover: String,
-)
